Version 1.0
This version of the library comes with a simple
Helpers.Firebase.FirebaseHelper object with 5 different functions.

Using the object is quite simply.

Initialize like so: Dim FbHelper As New
FirebaseHelper("FirebaseName","MyFirebaseSecretOrAuthTokenHere")

Then you can use any of the 5 methods. For example:
FbHelper.PushData("/users/3240/name","{""first"":""Chuck"",""last"":""Norris""}")

In the above case, the name of the newly pushed child node is returned.
Nothing is returned in the case of a failure. The other methods return
an enumeration member, or in the case of the ReadData function, the data
returned (or Nothing on an error).